+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 1.2
+- Added void printers to void seeds (configurable)
+    - Void seeds have ~50% chance to include void printers (configurable)
+    - Void printers have a 60% chance to be white, 36% to be green and 4% to be red (configurable)
+    - The credit cost is also configurable
 - Added white Cauldrons to the Bazaar (configurable).
 - Added yellow Cauldrons
     - They can be found in the Bazaar (configurable).
@@ -7,7 +11,6 @@
 - Encrusted Key caches now drop scrap by default (configurable).
 - Newt Altar now uses white items as the activation cost instead of lunar coins by default (configurable).
 - Lunar Seer (dream) now uses white items as the activation cost instead of lunar coins (configurable).
-
 
 ## 1.1.2
 - Updated documentation